Fixed input size type from float to int.

+++ convert.py
@@ -84,8 +84,8 @@ def _main(args):
     cfg_parser = configparser.ConfigParser()
     cfg_parser.read_file(unique_config_file)
 
-    width = float(cfg_parser["net_0"]["width"])
-    height = float(cfg_parser["net_0"]["height"])
+    width = int(cfg_parser["net_0"]["width"])
+    height = int(cfg_parser["net_0"]["height"])
 
     print('Creating Keras model.')
     input_layer = Input(shape=(height, width, 3))
